+++ client/src/components/QuickLink/index.js
@@ -1,4 +1,5 @@
 import React, { Component } from "react"
+import PropTypes from "prop-types";
 import {
 	RaisedButton
 } from "material-ui"
@@ -6,13 +7,11 @@ import { Link } from "react-router-dom"
 
 import styles from "./styles.scss"
 
-import { aNumber } from "./utils"
-
 // add images and finish layout of 2x2 for each cta
-export default function(routeObject) {
+export default function QuickLink(routeObject){
 	let {img, route} = routeObject.route
 	return (
-		<div className="dashboard-cta col-md-12">
+		<div className="dashboard-cta col-md-3">
 			<Link to={`/dashboard/${route}`}>
 		    	<RaisedButton className="">
 		    		{route}
@@ -21,3 +20,14 @@ export default function(routeObject) {
     	</div>
 	)
 }
+
+QuickLink.defaultProps = {
+  routeObject: {
+  	img: '',
+  	route: "test"
+  }
+};
+
+QuickLink.propTypes = {
+  routeObject: PropTypes.object
+};