+++ client/src/containers/dashboard.js
@@ -1,17 +1,17 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
 
-import ListItem from "../components/leaderboardItem";
+import QuickLink from "../components/QuickLink";
 import { dashboardRoutes } from "../constants";
 
 export default class Dashboard extends Component {
-	renderListItem(routes) {
+	renderQuickLink(routes) {
 		return routes.map((route, index) => {
-			return <ListItem key={index} route={route} />;
+			return <QuickLink key={index} route={route} />;
 		});
 	}
 
 	render() {
-		return <div>{this.renderListItem(dashboardRoutes)}</div>;
+		return <div>{this.renderQuickLink(dashboardRoutes)}</div>;
 	}
 }